requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
gherkin-official==29.0.0
idna==3.11
iniconfig==2.3.0
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
parse==1.20.2
parse_type==0.6.6
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.1
pytest-bdd==8.1.0
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
